requirements.txt
requests
fire
pyyaml
